refactor: 중복 메소드 추출
